Extract psi calculation routines into separate model file
